requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
httpx>=0.27.0
aiosqlite>=0.20.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
duckduckgo-search>=6.1.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
trafilatura>=1.9.0
yfinance>=0.2.40
pandas>=2.0.0
tenacity>=8.2.3
orjson>=3.10.0
